\name{AllfilesIEObyPathPattern}
\alias{AllfilesIEObyPathPattern}
\title{ Read multiple IEO files}
\description{\code{AllfilesIEObyPathPattern} reads multiple daily values file from IEO (Instituto Español de Oceanografia, Spain).
}
\usage{
AllfilesIEObyPathPattern(mypath, mypattern)
}
\arguments{
  \item{mypath}{ Path to search for files}
  \item{mypattern}{ Pattern to search for file names}
}
\value{
   A list with the read data.
}
\author{Jose Gama}
\keyword{programming}
