\name{AllReadUSGS2timeseries}
\alias{AllReadUSGS2timeseries}
\title{ Read multiple USGS files}
\description{\code{AllReadUSGS2timeseries} reads multiple USGS discharge files.
}
\usage{
AllReadUSGS2timeseries()
}
\arguments{
  none
}
\value{
   A list with the read data.
}
\author{Jose Gama}
\keyword{programming}
