\name{annual.precipitation.totals.Madison}
\alias{annual.precipitation.totals.Madison}
\title{Annual precipitation totals in inches Madison (Wisconsin) }
\usage{annual.precipitation.totals.Madison}
\description{\code{annual.precipitation.totals.Madison} contains the annual precipitation totals in inches from Madison (Wisconsin) - courtesy of Madison Climate Page/State Climatology Homepage.
}
\format{
  This data frame contains 2 columns (year and precipitationInches) 
and 135 rows of data.
}
\source{
Madison Climate Page/State Climatology Homepage
Annual precipitation totals in inches
\url{http://www.aos.wisc.edu/~sco/clim-history/stations/msn/msn-pcpn.html}
}
\references{
Madison Climate Page/State Climatology Homepage
Annual precipitation totals in inches
\url{http://www.aos.wisc.edu/~sco/clim-history/stations/msn/msn-pcpn.html}
}
\author{Jose Gama}
\examples{
data(annual.precipitation.totals.Madison)
str(annual.precipitation.totals.Madison)
}
\keyword{datasets}
