\name{Buishand.Critical.Values}
\alias{Buishand.Critical.Values}
\title{Critical values for the Buishand test}
\usage{Buishand.Critical.Values}
\description{\code{Buishand.Critical.Values} contains the critical 
values for the Buishand test.
}
\format{
  This data frame contains 6 columns (Q90%, Q95%, Q99%, R90%, R95% and R99%) 
and 7 rows (n=10, 20, 30, 40, 50, 100 and Inf) of data.
}
\source{
Buishand TA., 1981
The analysis of homogeneity of long-term rainfall records in the Netherlands.
KNMI Scientific Report WR 81-7, De Bilt, The Netherlands.
}
\references{
Buishand TA., 1981
The analysis of homogeneity of long-term rainfall records in the Netherlands.
KNMI Scientific Report WR 81-7, De Bilt, The Netherlands.
}
\author{Jose Gama}
\examples{
data(Buishand.Critical.Values)
Buishand.Critical.Values
}
\keyword{datasets}
