\name{ReadUSGS2timeseries}
\alias{ReadUSGS2timeseries}
\title{ Read a a USGS discharge file}
\description{\code{ReadUSGS2timeseries} reads a .part file which is a USGS discharge file.
}
\usage{
ReadUSGS2timeseries(filename)
}
\arguments{
  \item{filename}{ File name}
}
\value{
   A matrix with the read data.
}
\author{Jose Gama}
\keyword{programming}
