\name{IsLeapYear}
\alias{IsLeapYear}
\title{ Determine if a year is a leap year}
\description{\code{IsLeapYear} returns TRUE for a leap year.
}
\usage{
IsLeapYear(intYear)
}
\arguments{
  \item{intYear}{ year}
}
\value{
  logical, TRUE = leap year.
}
\author{Jose Gama}
\examples{
IsLeapYear(2010)
IsLeapYear(2011)
IsLeapYear(2012)
IsLeapYear(2013)
}
\keyword{programming}
