\name{Pettitt.Critical.Values}
\alias{Pettitt.Critical.Values}
\title{Critical values for the Pettitt test}
\usage{Pettitt.Critical.Values}
\description{\code{Pettitt.Critical.Values} contains the critical 
values for the Pettitt test.
}
\format{
  This data frame contains 2 columns (95% and 99%) 
and 6 rows (n=10, 20, 30, 40, 50, 100) of data.
}
\source{
Buishand TA., 1981
The analysis of homogeneity of long-term rainfall records in the Netherlands.
KNMI Scientific Report WR 81-7, De Bilt, The Netherlands.
}
\references{
Buishand TA., 1981
The analysis of homogeneity of long-term rainfall records in the Netherlands.
KNMI Scientific Report WR 81-7, De Bilt, The Netherlands.
}
\author{Jose Gama}
\examples{
data(Pettitt.Critical.Values)
Pettitt.Critical.Values
}
\keyword{datasets}
