\name{Rank.Von.Neumann.Ratio.Test.Critical.Values}
\alias{Rank.Von.Neumann.Ratio.Test.Critical.Values}
\title{Critical values for the Rank Von-Neumann Ratio Test}
\usage{Rank.Von.Neumann.Ratio.Test.Critical.Values}
\description{\code{Rank.Von.Neumann.Ratio.Test.Critical.Values} contains the critical 
values for the Rank Von-Neumann Ratio Test.
}
\format{
  This data frame contains 6 columns (Q90%, Q95%, Q99%, R90%, R95% and R99%) 
and 7 rows (n=10, 20, 30, 40, 50, 100 and Inf) of data.
}
\source{
Robert Bartels, 1982
The Rank Version of von Neumann's Ratio Test for Randomness
Journal of the American Statistical Association
Vol. 77, No. 377 pp. 40-46
\url{http://www.jstor.org/stable/2287767}
}
\references{
Robert Bartels, 1982
The Rank Version of von Neumann's Ratio Test for Randomness
Journal of the American Statistical Association
Vol. 77, No. 377 pp. 40-46
\url{http://www.jstor.org/stable/2287767}
}
\author{Jose Gama}
\examples{
data(Rank.Von.Neumann.Ratio.Test.Critical.Values)
Rank.Von.Neumann.Ratio.Test.Critical.Values
}
\keyword{datasets}
