\name{ReadClimexpKnmiNL12month}
\alias{ReadClimexpKnmiNL12month}
\title{ Read monthly values KNMI Climate Explorer files}
\description{\code{ReadClimexpKnmiNL12month} reads a monthly values file in climexp.knmi.nl format.
}
\usage{
ReadClimexpKnmiNL12month(filename)
}
\arguments{
  \item{filename}{ File name}
}
\value{
   A dataframe with the read data.
}
\author{Jose Gama}
\keyword{programming}
