\name{ReadClimexpKnmiNL}
\alias{ReadClimexpKnmiNL}
\title{ Read KNMI Climate Explorer files}
\description{\code{ReadClimexpKnmiNL} reads KNMI Climate Explorer files.
}
\usage{
ReadClimexpKnmiNL(filename)
}
\arguments{
  \item{filename}{ File name}
}
\value{
   A dataframe with the read data.
}
\author{Jose Gama}
\keyword{programming}
