\name{ReadECAdata}
\alias{ReadECAdata}
\alias{ReadECAdataIndexDTR}
\title{ Read data from a ECA data file }
\description{\code{ReadECAdata} reads data from a file downloaded from the European Climate Assessment & Dataset project.
}
\usage{
ReadECAdata(filename)
}
\arguments{
  \item{filename}{ File name}
}
\value{
   A matrix with the read data.
}
\author{Jose Gama}
\keyword{programming}
