\name{ReadHERTTAdailyCSV}
\alias{ReadHERTTAdailyCSV}
\title{ Read HERTTA files}
\description{\code{ReadHERTTAdailyCSV} reads a file from Finland's Environmental centre (HERTTA) into a dataframe.
}
\usage{
ReadHERTTAdailyCSV(filename)
}
\arguments{
  \item{filename}{ File name}
}
\value{
   A dataframe with the read data.
}
\author{Jose Gama}
\keyword{programming}
