\name{ReadHERTTAdmyQdaily}
\alias{ReadHERTTAdmyQdaily}
\title{ Read HERTTA month+year+Q files}
\description{\code{ReadHERTTAdmyQdaily} reads a month+year+Q file from Finland's Environmental centre (HERTTA) into a dataframe.
}
\usage{
ReadHERTTAdmyQdaily(filename)
}
\arguments{
  \item{filename}{ File name}
}
\value{
   A dataframe with the read data.
}
\author{Jose Gama}
\keyword{programming}
