\name{ReadMeteoSwiss}
\alias{ReadMeteoSwiss}
\title{ Read MeteoSwiss files}
\description{\code{ReadMeteoSwiss} reads a daily values file from MeteoSwiss.
}
\usage{
ReadMeteoSwiss(filename)
}
\arguments{
  \item{filename}{ File name}
}
\value{
   A dataframe with the read data.
}
\author{Jose Gama}
\keyword{programming}
