\name{ReadNdp040stationDat}
\alias{ReadNdp040stationDat}
\alias{ReadNdp040stationData}
\alias{ReadNdp040stationHistory}
\alias{ReadNdp040stationInventory}
\title{ Read data from a CDIAC data file }
\description{\code{ReadNdp040stationDat} reads data from a file downloaded 
from the Carbon Dioxide Information Analysis Center (CDIAC).
Daily Temperature and Precipitation Data for 223 Former-USSR Stations.
}
\usage{
ReadNdp040stationDat(filename)
}
\arguments{
  \item{filename}{ File name}
}
\value{
   A matrix with the read data.
}
\author{Jose Gama}
\keyword{programming}
