\name{ReadPSMSLmonthly}
\alias{ReadPSMSLmonthly}
\title{ Read PSMSL files}
\description{\code{ReadPSMSLmonthly} reads a file from the Permanent Service for Mean Sea Level (PSMSL) into a dataframe.
}
\usage{
ReadPSMSLmonthly(filename)
}
\arguments{
  \item{filename}{ File name}
}
\value{
   A dataframe with the read data.
}
\author{Jose Gama}
\keyword{programming}
