\name{ReadUSHCN}
\alias{ReadUSHCN}
\title{ Read USHCN files}
\description{\code{ReadUSHCN} reads a file from the United States Historical Climatology Network (USHCN) into a dataframe.
}
\usage{
ReadUSHCN(filename)
}
\arguments{
  \item{filename}{ File name}
}
\value{
   A dataframe with the read data.
}
\author{Jose Gama}
\keyword{programming}
