\name{ReadYearAnd12MonthsData}
\alias{ReadYearAnd12MonthsData}
\title{ Read a TAB delimited data file with year and months in columns}
\description{\code{ReadYearAnd12MonthsData} reads a monthly values file with the data as y + 12 months in 1 line.
}
\usage{
ReadYearAnd12MonthsData(filename)
}
\arguments{
  \item{filename}{ File name}
}
\value{
   A matrix with the read data.
}
\author{Jose Gama}
\keyword{programming}
