\name{SNHT.Critical.Values}
\alias{SNHT.Critical.Values}
\alias{SNHT.Critical.Values.Alexandersson1997}
\title{Critical values for the Pettitt test}
\usage{SNHT.Critical.Values}
\description{\code{SNHT.Critical.Values} contains the critical 
values for the SNHT test from Khaliq and Ouarda, 2007. \code{SNHT.Critical.Values.Alexandersson1997} has
the data from Alexandersson 1997, which is used in AnClim.
}
\format{
  This data frame contains 2 columns (90.00%, 92.00%, 94.00%, 95.00%, 97.50%, 99.00%) 
and 108 rows (n=10, 12, 14, 16, ..., 15000, 20000, 50000) of data.
}
\source{
Khaliq, M. N., Ouarda, T. B. M. J. 2007
On the critical values of the standard normal homogeneity test (SNHT)
International Journal of Climatology
John Wiley & Sons, Ltd.
\url{http://dx.doi.org/10.1002/joc.1438}
}
\references{
Khaliq, M. N., Ouarda, T. B. M. J. 2007
On the critical values of the standard normal homogeneity test (SNHT)
International Journal of Climatology
John Wiley & Sons, Ltd.
\url{http://dx.doi.org/10.1002/joc.1438}
}
\author{Jose Gama}
\examples{
data(SNHT.Critical.Values)
SNHT.Critical.Values
}
\keyword{datasets}
