\name{SNHT.TwoDifferentSDs.Critical.Values}
\alias{SNHT.TwoDifferentSDs.Critical.Values}
\title{Critical values for the standard normal homogeneity test (SNHT) for two different SDs}
\usage{SNHT.TwoDifferentSDs.Critical.Values}
\description{\code{SNHT.TwoDifferentSDs.Critical.Values} contains the critical 
values for the SNHT test for 2 different SDs from Alexandersson and Moberg (1997).
}
\format{
  This data frame contains 2 columns (90.00%, 92.00%, 94.00%, 95.00%, 97.50%, 99.00%) 
  and 108 rows (n=10, 12, 14, 16, ..., 15000, 20000, 50000) of data.
}
\source{
ALEXANDERSSON, H. and MOBERG, A. (1997), HOMOGENIZATION OF SWEDISH TEMPERATURE DATA.
PART I: HOMOGENEITY TEST FOR LINEAR TRENDS. Int. J. Climatol., 17: 25–34. 
doi: 10.1002/(SICI)1097-0088(199701)17:1<25::AID-JOC103>3.0.CO;2-J

Tuomenvirta, H., 2002: Homogeneity testing and adjustment of climatic time series in Finland. Geophysica, 38(1-2), 15-41.
}
\references{
ALEXANDERSSON, H. and MOBERG, A. (1997), HOMOGENIZATION OF SWEDISH TEMPERATURE DATA.
PART I: HOMOGENEITY TEST FOR LINEAR TRENDS. Int. J. Climatol., 17: 25–34. 
doi: 10.1002/(SICI)1097-0088(199701)17:1<25::AID-JOC103>3.0.CO;2-J

Tuomenvirta, H., 2002: Homogeneity testing and adjustment of climatic time series in Finland. Geophysica, 38(1-2), 15-41.
}
\author{Jose Gama}
\examples{
data(SNHT.TwoDifferentSDs.Critical.Values)
SNHT.TwoDifferentSDs.Critical.Values
}
\keyword{datasets}
