\name{SNHT.TwoShifts.Critical.Values}
\alias{SNHT.TwoShifts.Critical.Values}
\title{Critical values for the standard normal homogeneity test (SNHT) for two shifts}
\usage{SNHT.TwoShifts.Critical.Values}
\description{\code{SNHT.TwoShifts.Critical.Values} contains the critical 
values for the SNHT test for 2 shifts from Alexandersson (1995).
}
\format{
  This data frame contains 2 columns (90.00%, 92.00%, 94.00%, 95.00%, 97.50%, 99.00%) 
  and 108 rows (n=10, 12, 14, 16, ..., 15000, 20000, 50000) of data.
}
\source{
Alexandersson, H., 1995. Homogeneity testing, multiple breaks and trends.
Proceedings of the Sixth International Meeting on Statistical Climatology,
Galway, Ireland, 439-441.

Tuomenvirta, H., 2002: Homogeneity testing and adjustment of climatic time series in Finland. Geophysica, 38(1-2), 15-41.
}
\references{
Alexandersson, H., 1995. Homogeneity testing, multiple breaks and trends.
Proceedings of the Sixth International Meeting on Statistical Climatology,
Galway, Ireland, 439-441.

Tuomenvirta, H., 2002: Homogeneity testing and adjustment of climatic time series in Finland. Geophysica, 38(1-2), 15-41.
}
\author{Jose Gama}
\examples{
data(SNHT.TwoShifts.Critical.Values)
SNHT.TwoShifts.Critical.Values
}
\keyword{datasets}
