\name{Sqplot}
\alias{Sqplot}
\title{ Number of elements for a plot to be nearly square}
\description{\code{Sqplot} returns the number of elements for a plot to be nearly square.
}
\usage{
Sqplot(n)
}
\arguments{
  \item{n}{ number of elements}
}
\value{
  vector with number of rows and columns.
}
\author{Jose Gama}
\examples{
Sqplot(10)
}
\keyword{programming}
