\name{Von.Neumann.Ratio.Test.Critical.Values}
\alias{Von.Neumann.Ratio.Test.Critical.Values}
\title{Critical values for the Von Neumann ratio test}
\usage{Von.Neumann.Ratio.Test.Critical.Values}
\description{\code{Von.Neumann.Ratio.Test.Critical.Values} contains the critical 
values for the Von Neumann ratio test.
}
\format{
  This data frame contains 2 columns (1% and 5%) 
and 6 rows (n=20, 30, 40, 50, 70 and 100) of data.
}
\source{
Von Neumann J., 1941
Distribution of the ratio of the mean square successive difference to the variance.
Annals of Mathematical Statistics 13: 367–395.
}
\references{
Von Neumann J., 1941
Distribution of the ratio of the mean square successive difference to the variance.
Annals of Mathematical Statistics 13: 367–395.
}
\author{Jose Gama}
\examples{
data(Von.Neumann.Ratio.Test.Critical.Values)
Von.Neumann.Ratio.Test.Critical.Values
}
\keyword{datasets}
