\name{Worsley.Likelihood.Ratio.Test.Critical.Values}
\alias{Worsley.Likelihood.Ratio.Test.Critical.Values}
\title{Critical values for Worsley's likelihood ratio test}
\usage{Worsley.Likelihood.Ratio.Test.Critical.Values}
\description{\code{Worsley.Likelihood.Ratio.Test.Critical.Values} contains the critical 
values for the Worsley's likelihood ratio test.
}
\format{
  This data frame contains 6 columns (Q90%, Q95%, Q99%, R90%, R95% and R99%) 
and 7 rows (n=10, 20, 30, 40, 50, 100 and Inf) of data.
}
\source{
Worsley, K.J., 1979
On the likelihood ratio test for a shift in location of normal populations.
J. Am. Stat. Assoc., 74: 365--367.
}
\references{
Worsley, K.J., 1979
On the likelihood ratio test for a shift in location of normal populations.
J. Am. Stat. Assoc., 74: 365--367.
}
\author{Jose Gama}
\examples{
data(Worsley.Likelihood.Ratio.Test.Critical.Values)
Worsley.Likelihood.Ratio.Test.Critical.Values
}
\keyword{datasets}
