\name{yearly.average.temperature.Turin.Milan}
\alias{yearly.average.temperature.Turin.Milan}
\title{Yearly average temperature in Turin, Milan}
\usage{yearly.average.temperature.Turin.Milan}
\description{\code{yearly.average.temperature.Turin.Milan} contains the yearly average temperature in Turin, Milan.
}
\format{
  This data frame contains 6 columns (Q90%, Q95%, Q99%, R90%, R95% and R99%) 
and 7 rows (n=10, 20, 30, 40, 50, 100 and Inf) of data.
}
\source{
Historical Climatology Group, 2005
Training session on homogenisation methods Bologna 17-18 May, 2005
}
\references{
Historical Climatology Group, 2005
Training session on homogenisation methods Bologna 17-18 May, 2005
}
\author{Jose Gama}
\examples{
data(yearly.average.temperature.Turin.Milan)
str(yearly.average.temperature.Turin.Milan)
}
\keyword{datasets}
