\name{Worsley.Likelihood.Ratio.Test.Critical.Values}
\alias{Worsley.Likelihood.Ratio.Test.Critical.Values}
\title{Critical values for Worsley's likelihood ratio test}
\usage{Worsley.Likelihood.Ratio.Test.Critical.Values}
\description{\code{Worsley.Likelihood.Ratio.Test.Critical.Values} contains the critical 
values for the Worsley's likelihood ratio test.
}
\format{
  This data frame contains 3 columns (alpha10%, alpha5%, alpha1%) 
and 9 rows (n=10, 15, 20, 25, 30, 35, 40, 45, 50) of data.
}
\source{
Worsley, K.J., 1979
On the likelihood ratio test for a shift in location of normal populations.
J. Am. Stat. Assoc., 74: 365--367.
}
\references{
Worsley, K.J., 1979
On the likelihood ratio test for a shift in location of normal populations.
J. Am. Stat. Assoc., 74: 365--367.
}
\author{Jose Gama}
\examples{
data(Worsley.Likelihood.Ratio.Test.Critical.Values)
Worsley.Likelihood.Ratio.Test.Critical.Values
}
\keyword{datasets}
